Code_Review_Bot/requirements.txt
streamlit
google-generativeai
langchain
langchain-community
langchain-google-genai
chromadb
pypdf
python-docx
python-dotenv